9_Azure_AI/1_azure-ml/0_basics_old_sdk/pyproject.toml
[tool.poetry]
name = "project1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
azureml-core = "^1.52.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
